+++ docs/scripts/services/calculatedDateHistory.service.js
@@ -35,7 +35,6 @@
 			reset,
 			setOriginal,
 			setCalculated,
-			setFinal,
 			addFerie,
 			addSunday,
 			addSaturday