+++ docs/release/x-days-later.js
@@ -299,7 +299,6 @@
 			reset: reset,
 			setOriginal: setOriginal,
 			setCalculated: setCalculated,
-			setFinal: setFinal,
 			addFerie: addFerie,
 			addSunday: addSunday,
 			addSaturday: addSaturday
@@ -418,8 +417,6 @@
 			isSaturday: isSaturday,
 			isDay: isDay,
 			isWeekend: isWeekend,
-			setNextDayAfterWeekend: setNextDayAfterWeekend,
-			weekendAndExceptionsStuff: weekendAndExceptionsStuff,
 			readable: readable,
 			toTimestamp: toTimestamp
 		};
@@ -505,55 +502,6 @@
 
 		function isWeekend($date) {
 			return methods.isSunday($date) || methods.isSaturday($date);
-		}
-
-		function setNextDayAfterWeekend($date) {
-			if (methods.isSunday($date)) {
-				return methods.add($date, data.one, 'days');
-			}
-			return methods.add($date, data.two, 'days');
-		}
-
-		function weekendAndExceptionsStuff($date) {
-			var date = $date;
-
-			// Check if this is the weekend
-			if (methods.isWeekend(date)) {
-				logService.service(data.service, 'isWeekend');
-				var weekend = {
-					dateBefore: methods.toTimestamp(date),
-					dateAfter: null,
-					type: methods.isSunday(date) ? 'sunday' : 'saturday'
-				};
-
-				// Set the next monday
-				date = methods.setNextDayAfterWeekend(date);
-				logService.service(data.service, 'new date after weekend: ' + methods.readable(date));
-				weekend.dateAfter = methods.toTimestamp(date);
-				calculatedDateHistoryService.addWeekend(weekend);
-			}
-
-			// Check if this is an fr exception
-			if (date.isFerie()) {
-				logService.service(data.service, 'isFerie');
-				var ferie = {
-					dateBefore: methods.toTimestamp(date),
-					dateAfter: null,
-					ferie: date.getFerie()
-				};
-
-				// Add one day
-				date = methods.add(date, data.one, 'days');
-				logService.service(data.service, 'new date after exception day: ' + methods.readable(date));
-				ferie.dateAfter = methods.toTimestamp(date);
-				calculatedDateHistoryService.addFerie(ferie);
-
-				// Loop again
-				date = weekendAndExceptionsStuff(date);
-			}
-			logService.service(data.service, 'final new date: ' + methods.readable(date));
-			calculatedDateHistoryService.setFinal(methods.toTimestamp(date));
-			return methods.toTimestamp(date);
 		}
 
 		function readable($date) {
@@ -699,7 +647,7 @@ function safeApply(scope, fn) {
 	config.$inject = [];
 
 	function config() {
-		console.info('x-days-later version: 0.11.0');
+		console.info('x-days-later version: 0.11.1');
 	}
 })(window.angular);